requirements.txt
requests
pytest
hypothesis
pyyaml
jsonschema
mypy